project3-ml/requirements.txt
pyspark
kafka-python
python-dotenv
influxdb